refactor: split off StackTest.Repl
